+++ src/osax/loader.m
@@ -1,3 +1,0 @@
-#ifdef __x86_64__
-#include "x86_64/loader.m"
-#endif