+++ src/sa.h
@@ -27,12 +27,7 @@ bool scripting_addition_swap_window_order(uint32_t a_wid, uint32_t b_wid);
 
 extern bool workspace_is_macos_monterey(void);
 extern bool workspace_is_macos_bigsur(void);
-extern bool workspace_is_macos_highsierra(void);
 
-extern unsigned char __src_osax_mach_bootstrap[];
-extern unsigned int __src_osax_mach_bootstrap_len;
-extern unsigned char __src_osax_loader[];
-extern unsigned int __src_osax_loader_len;
 extern unsigned char __src_osax_payload[];
 extern unsigned int __src_osax_payload_len;
 extern unsigned char __src_osax_mach_loader[];