+++ src/workspace.m
@@ -134,12 +134,6 @@ static int workspace_is_macos_major(long major_version)
     return version.majorVersion == major_version;
 }
 
-static int workspace_is_macos_minor(long major_version, long minor_version)
-{
-    NSOperatingSystemVersion version = [[NSProcessInfo processInfo] operatingSystemVersion];
-    return version.majorVersion == major_version && version.minorVersion == minor_version;
-}
-
 bool workspace_is_macos_monterey(void)
 {
     if (_workspace_is_macos_version[0] == -1) {
@@ -154,30 +148,6 @@ bool workspace_is_macos_bigsur(void)
         _workspace_is_macos_version[1] = workspace_is_macos_major(11);
     }
     return _workspace_is_macos_version[1];
-}
-
-bool workspace_is_macos_catalina(void)
-{
-    if (_workspace_is_macos_version[2] == -1) {
-        _workspace_is_macos_version[2] = workspace_is_macos_minor(10, 15);
-    }
-    return _workspace_is_macos_version[2];
-}
-
-bool workspace_is_macos_mojave(void)
-{
-    if (_workspace_is_macos_version[3] == -1) {
-        _workspace_is_macos_version[3] = workspace_is_macos_minor(10, 14);
-    }
-    return _workspace_is_macos_version[3];
-}
-
-bool workspace_is_macos_highsierra(void)
-{
-    if (_workspace_is_macos_version[4] == -1) {
-        _workspace_is_macos_version[4] = workspace_is_macos_minor(10, 13);
-    }
-    return _workspace_is_macos_version[4];
 }
 
 @implementation workspace_context